+++ packages/server/src/api/controllers/Sales/SalesReceipts.ts
@@ -139,7 +139,7 @@ export default class SalesReceiptsController extends BaseController {
       check('entries.*.id').optional({ nullable: true }).isNumeric().toInt(),
       check('entries.*.index').exists().isNumeric().toInt(),
       check('entries.*.item_id').exists().isNumeric().toInt(),
-      check('entries.*.quantity').exists().isNumeric().toInt(),
+      check('entries.*.quantity').exists().isNumeric().toFloat(),
       check('entries.*.rate').exists().isNumeric().toFloat(),
       check('entries.*.discount')
         .optional({ nullable: true })
